Fixed some small issues and added some small features

+++ src/Controller/ModifyProductFormController.java
@@ -142,7 +142,11 @@ public class ModifyProductFormController implements Initializable {
             nothingSelectedErrorMsg();
             return;
         }
-        tempProd.deleteAssociatedPart(p);
+        else if(confirmDelete(p.getName()))
+        {
+            tempProd.deleteAssociatedPart(p);
+        }
+
     }
 
     /**
@@ -159,16 +163,12 @@ public class ModifyProductFormController implements Initializable {
             }
             if (idInput.getText().isBlank() || nameInput.getText().isBlank() || invInput.getText().isBlank() || priceInput.getText().isBlank() || maxInput.getText().isBlank() || minInput.getText().isBlank())
             {
-                Alert alert = new Alert(Alert.AlertType.ERROR);
-                alert.setTitle("Empty Input");
-                alert.setHeaderText("Empty Input");
-                alert.setContentText("Please make sure there is information in all fields");
-                alert.showAndWait();
+                emptySearchErrorMsg();
                 return;
             }
             if(Double.parseDouble(priceInput.getText()) < totalPrice)
             {
-                nothingSelectedErrorMsg();
+                wrongPriceErrorMsg();
                 return;
             }
             tempProd.setID(Integer.parseInt(idInput.getText()));
@@ -177,7 +177,11 @@ public class ModifyProductFormController implements Initializable {
             tempProd.setStock(Integer.parseInt(invInput.getText()));
             tempProd.setMax( Integer.parseInt(maxInput.getText()));
             tempProd.setMin(Integer.parseInt(minInput.getText()));
-
+            if(!checkInvMinMax(tempProd.getStock(),tempProd.getMin(),tempProd.getMax()))
+            {
+                invMinMaxErrorMsg();
+                return;
+            }
             inv.deleteProduct(prod);
             inv.addProduct(tempProd);
             closeForm();
@@ -258,6 +262,46 @@ public class ModifyProductFormController implements Initializable {
         allPartsTbl.refresh();
     }
 
+    /**
+     * Tells if Inventory, Max, and min have values that work.
+     * @param inv amount of inventory
+     * @param min the minimum amount
+     * @param max the maximum amount
+     * @return wether or not max is greater than or equal to inventory and inventory is greater than or equal to minimum
+     */
+    private boolean checkInvMinMax(int inv, int min, int max)
+    {
+        if(max >= inv && inv >= min)
+        {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * Brings up a dialog to confirm if the user wants a part deleted
+     * @param name the name of the part the user wants deleted
+     * @return true if the user choses yes. false if not.
+     */
+    private boolean confirmDelete(String name)
+    {
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Confirm removal");
+        alert.setHeaderText("Confirm removal");
+        alert.setContentText("Are you sure you want to delete " + name + "?");
+        ButtonType yesBtn = new ButtonType("Yes", ButtonBar.ButtonData.YES);
+        ButtonType noBtn = new ButtonType("No", ButtonBar.ButtonData.NO);
+
+        alert.getButtonTypes().setAll(yesBtn, noBtn);
+
+        ButtonType anwser = alert.showAndWait().orElse(ButtonType.NO);
+        if(anwser == yesBtn)
+        {
+            return true;
+        }
+        return false;
+    }
+
     //==================================================================
     //Error Message Methods
     //==================================================================
@@ -297,6 +341,31 @@ public class ModifyProductFormController implements Initializable {
         alert.setTitle("Nothing Selected");
         alert.setHeaderText("Nothing Selected");
         alert.setContentText("Please select a part to Modify");
+        alert.showAndWait();
+    }
+
+    /**
+     * Displays an error message for incorrect input
+     */
+    private void invMinMaxErrorMsg()
+    {
+        Alert alert = new Alert(Alert.AlertType.ERROR);
+        alert.setTitle("Invalid Inventory, Min, or Max");
+        alert.setHeaderText("Invalid Inventory, Min, or Max");
+        alert.setContentText("Max must be greater than Min and inventory must be in between or equal to them");
+
+        alert.showAndWait();
+    }
+
+    /**
+     * Displays an error message for a wrong price
+     */
+    private void wrongPriceErrorMsg()
+    {
+        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.setTitle("Incorrect Price");
+        alert.setHeaderText("Incorrect Price");
+        alert.setContentText("Please make sure your total price is more than the sum of your part prices.");
         alert.showAndWait();
     }
 }
